requirements.txt
requests
html2text
tqdm
python-dotenv
qdrant-client
google-cloud-aiplatform
vertexai
tiktoken
